% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominosa.R
\name{solve_dominosa}
\alias{solve_dominosa}
\title{Solve Dominosa puzzle}
\usage{
solve_dominosa(pips)
}
\arguments{
\item{pips}{A matrix of non-negative integer pip values,
or a string with rows of integers separated by \code{"/"}.
For a double-\code{n} domino set the grid must have \code{(n + 1)(n + 2)} cells
and each pip value \code{0}:\code{n} must appear exactly \code{n + 2} times.}
}
\value{
A list with components:
\describe{
\item{\code{domino_ids}}{An integer matrix of the same dimensions as \code{pips}
where cells belonging to the same domino share the same integer value.}
\item{\code{pips}}{The input \code{pips} matrix.}
\item{\code{ppn}}{A string of Portable Piecepack Notation for the solution.}
}
}
\description{
Solves a Dominosa (Domino Solitaire) puzzle given a grid of pip values.
}
\examples{
pips <- matrix(c(0, 1, 0, 1, 3,
                 3, 1, 0, 2, 2,
                 3, 2, 1, 0, 3,
                 2, 3, 1, 0, 2), nrow = 4L, byrow = TRUE)
s <- solve_dominosa(pips)
if (rlang::is_installed(c("piecepackr", "ppn"))) {
  g <- ppn::read_ppn(textConnection(s$ppn))[[1]]
  envir <- piecepackr::game_systems(round = TRUE)
  ppn::plot_move(g, open_device = FALSE, annotate = FALSE, envir = envir, scale = 0.95)
}
}
\seealso{
\url{https://www.solitairelaboratory.com/puzzlelaboratory/DominoGG.html} and
\url{https://puzzlebreaks.com/dominosa/} for more information about Dominosa.
}
