% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fujisan.R
\name{solve_fujisan}
\alias{solve_fujisan}
\title{Solve Fujisan game}
\usage{
solve_fujisan(
  coins = random_fujisan_coins(),
  dice = random_dice() - 1,
  pawns = "S12M/A12C"
)
}
\arguments{
\item{coins}{A vector or matrix of Fujisan coin layout.  Default is a random layout.}

\item{dice}{A vector of Fujisan dice layout.  Default is random dice.  Usually not needed.}

\item{pawns}{A FEN-like string of the pawns position.  The four pawns \code{S}, \code{M}, \code{C}, and
\code{A} are represented by their letter and empty spaces by a number.  The two rows are
separated by \code{/}.  Default is \code{"S12M/A12C"} (all pawns at starting corners).}
}
\value{
A list with solution of Fujisan solution, its length, coin layout, dice (if needed),
pawn position string, and portable piecepack notation.
}
\description{
Solves a game of Fujisan (if possible).
}
\examples{
 puzzle2 <- matrix(c(4, 4, 4, 5, 2, 0, 2, 4, 0, 3, 1, 1,
                     1, 2, 5, 3, 3, 5, 3, 2, 5, 1, 0, 0), nrow = 2, byrow = TRUE)
 s <- solve_fujisan(coins = puzzle2)
 print(s$shortest_distance) # minimum number of moves to solve game
}
